BPI及Content  查詢/新增/修改/刪除

+++ Currency.API.Domain/IRepositories/Bpi/Model/InsertBpiInput.cs
@@ -7,6 +7,6 @@ namespace Currency.API.Domain.IRepositories.Bpi.Model
 		public string CurrencyCode { get; set; } = null!;
 		public string Symbol { get; set; } = null!;
 		public string Rate { get; set; } = null!;
-		public double RateFloat { get; set; }
+		public decimal RateFloat { get; set; }
 	}
 }
